+++ compiler-core/src/analyse.rs
@@ -1255,6 +1255,7 @@ impl<'a, A> ModuleAnalyzer<'a, A> {
         let Accessors {
             shared_accessors,
             variant_specific_accessors,
+            positional_accessors,
         } = custom_type_accessors(&constructors_data)?;
 
         let map = AccessorsMap {
@@ -1268,6 +1269,7 @@ impl<'a, A> ModuleAnalyzer<'a, A> {
             // `return_type_constructor` below rather than looking it up twice.
             type_: type_.clone(),
             variant_specific_accessors,
+            variant_positional_accessors: positional_accessors,
         };
         environment.insert_accessors(name.clone(), map);
 
@@ -1955,6 +1957,7 @@ fn assert_unique_name(
 struct Accessors {
     shared_accessors: HashMap<EcoString, RecordAccessor>,
     variant_specific_accessors: Vec<HashMap<EcoString, RecordAccessor>>,
+    positional_accessors: Vec<Vec<Arc<Type>>>,
 }
 
 fn custom_type_accessors(constructors: &[TypeValueConstructor]) -> Result<Accessors, Error> {
@@ -1966,33 +1969,36 @@ fn custom_type_accessors(constructors: &[TypeValueConstructor]) -> Result<Access
         let _ = shared_accessors.insert(accessor.label.clone(), accessor);
     }
 
-    let mut variant_specific_accessors: Vec<HashMap<EcoString, RecordAccessor>> =
-        Vec::with_capacity(constructors.len());
+    let mut variant_specific_accessors = Vec::with_capacity(constructors.len());
+    let mut positional_accessors = Vec::with_capacity(constructors.len());
 
     for constructor in constructors {
         let mut fields = HashMap::with_capacity(constructor.parameters.len());
+        let mut positional_fields = Vec::new();
 
         for (index, parameter) in constructor.parameters.iter().enumerate() {
-            let Some(label) = &parameter.label else {
-                continue;
-            };
-
-            let _ = fields.insert(
-                label.clone(),
-                RecordAccessor {
-                    index: index as u64,
-                    label: label.clone(),
-                    type_: parameter.type_.clone(),
-                    documentation: parameter.documentation.clone(),
-                },
-            );
+            if let Some(label) = &parameter.label {
+                _ = fields.insert(
+                    label.clone(),
+                    RecordAccessor {
+                        index: index as u64,
+                        label: label.clone(),
+                        type_: parameter.type_.clone(),
+                        documentation: parameter.documentation.clone(),
+                    },
+                );
+            } else {
+                positional_fields.push(parameter.type_.clone());
+            }
         }
         variant_specific_accessors.push(fields);
+        positional_accessors.push(positional_fields);
     }
 
     Ok(Accessors {
         shared_accessors,
         variant_specific_accessors,
+        positional_accessors,
     })
 }